Widen ac_ir_frame_t.durations to uint32_t
Logical protocol durations are conceptually unbounded microsecond
values and shouldn't be tied to the RMT hardware's 15-bit duration
field (that limit is already handled separately by the split logic
in ac_ir.c). uint16_t happened to be sufficient for the durations
seen so far, but caps at 65535us, which some protocol gaps could
plausibly exceed.

ac_ir.c requires no changes: it already reads frame->durations[j]
into a local uint32_t, so the split/pack logic is unaffected.

Protocol sources (gree.c, haier.c, etc.) that declare raw duration
arrays as uint16_t rawData[] must be updated to uint32_t rawData[]
to match the new field type -- no timing values change, this is a
mechanical type update.

+++ firmware/components/13_ac_control/src/ac_ir_frame.h
@@ -11,8 +11,17 @@
  * durations[2] = mark
  * durations[3] = space
  * ...
+ *
+ * These are logical, protocol-level durations in microseconds and
+ * are NOT constrained by the RMT hardware's 15-bit duration field.
+ * A single logical duration may be larger than the hardware can
+ * represent in one RMT symbol -- the generic TX encoder (ac_ir.c)
+ * is responsible for splitting such durations into multiple RMT
+ * half-durations of the same level. Protocol sources (gree.c,
+ * haier.c, etc.) do not need to know about that limit and should
+ * just express the real waveform timing here.
  */
 typedef struct {
-    uint16_t *durations;
+    uint32_t *durations;
     size_t count;
 } ac_ir_frame_t;
